removed some unneeded parts from ScreenControlUI import
