--- TouchFree/Assets/TouchFree/Scripts/UI/UIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39d0a2ff7ff1b0c4a8dc636bba3b294d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
